For å kunne gjenbruke MapComponent i observasjonskort, måtte vi gjøre om litt på modulinndeling. Flere av komponentene i MapModule er nå organisert i egne moduler.
