--- src/app/modules/map-image/map-image.component.html
@@ -1,2 +1,0 @@
-<div leaflet [leafletOptions]="options" (leafletMapReady)="onLeafletMapReady($event)">
-</div>
