--- src/app/modules/map-image/map-image.module.ts
@@ -1,20 +1,0 @@
-import { MapImageComponent } from './map-image.component';
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
-import { IonicModule } from '@ionic/angular';
-import { TranslateModule } from '@ngx-translate/core';
-import { LeafletModule } from '@asymmetrik/ngx-leaflet';
-
-@NgModule({
-  imports: [
-    CommonModule,
-    FormsModule,
-    IonicModule,
-    TranslateModule,
-    LeafletModule
-  ],
-  declarations: [MapImageComponent],
-  exports: [MapImageComponent]
-})
-export class MapImageModule {}
