+++ src/app/modules/map/components/map-component.module.ts
@@ -1,0 +1,11 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { MapControlsModule } from './map-controls/map-controls.module';
+import { MapComponent } from './map/map.component';
+
+@NgModule({
+  imports: [CommonModule, MapControlsModule],
+  declarations: [MapComponent],
+  exports: [MapComponent, MapControlsModule]
+})
+export class MapComponentModule {}
