+++ src/app/modules/map/components/map-image/map-image.component.html
@@ -1,0 +1,3 @@
+<!--div leaflet [leafletOptions]="options" (leafletMapReady)="onLeafletMapReady($event)">
+</div-->
+<app-map></app-map>
