+++ src/app/modules/map/components/map-image/map-image.component.spec.ts
