+++ src/app/modules/map/map.module.ts
@@ -1,22 +1,24 @@
+import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { MapComponent } from './components/map/map.component';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { RouterModule } from '@angular/router';
 import { LeafletModule } from '@asymmetrik/ngx-leaflet';
-import { MapControlsComponent } from './components/map-controls/map-controls.component';
-import { MapSearchComponent } from './components/map-controls/map-search/map-search.component';
+import { IonicModule } from '@ionic/angular';
+import { TranslateModule } from '@ngx-translate/core';
+import { AngularSvgIconModule } from 'angular-svg-icon';
+import { LeafletEdgeBufferModule } from 'ngx-leaflet-edgebuffer';
+import { MapCenterInfoComponent } from './components/map-center-info/map-center-info.component';
 import { FullscreenToggleComponent } from './components/map-controls/fullscreen-toggle/fullscreen-toggle.component';
 import { GpsCenterComponent } from './components/map-controls/gps-center/gps-center.component';
-import { MapCenterInfoComponent } from './components/map-center-info/map-center-info.component';
-import { ModalSearchPageModule } from './pages/modal-search/modal-search.module';
-import { LeafletEdgeBufferModule } from 'ngx-leaflet-edgebuffer';
-import { CommonModule } from '@angular/common';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { IonicModule } from '@ionic/angular';
-import { AngularSvgIconModule } from 'angular-svg-icon';
-import { TranslateModule } from '@ngx-translate/core';
-import { RouterModule } from '@angular/router';
+import { MapControlsComponent } from './components/map-controls/map-controls.component';
+import { MapSearchComponent } from './components/map-controls/map-search/map-search.component';
+import { MapImageComponent } from './components/map-image/map-image.component';
+import { MapImageModule } from './components/map-image/map-image.module';
+import { MapComponentModule } from './components/map-component.module';
 import { ModalMapImagePageModule } from './pages/modal-map-image/modal-map-image.module';
+import { ModalSearchPageModule } from './pages/modal-search/modal-search.module';
 import { SupportMapInfoPageModule } from './pages/support-map-info/support-map-info.module';
-import { MapImageModule } from '../map-image/map-image.module';
+import { MapControlsModule } from './components/map-controls/map-controls.module';
 
 @NgModule({
   imports: [
@@ -30,24 +32,16 @@ import { MapImageModule } from '../map-image/map-image.module';
     LeafletModule,
     LeafletEdgeBufferModule,
     ModalSearchPageModule,
+    MapComponentModule,
     ModalMapImagePageModule,
     SupportMapInfoPageModule,
     MapImageModule
   ],
-  declarations: [
-    MapComponent,
-    MapControlsComponent,
-    MapSearchComponent,
-    FullscreenToggleComponent,
-    GpsCenterComponent,
-    MapCenterInfoComponent
-  ],
+  declarations: [MapCenterInfoComponent],
   exports: [
-    MapComponent,
-    MapControlsComponent,
-    MapSearchComponent,
-    FullscreenToggleComponent,
-    GpsCenterComponent,
+    MapComponentModule,
+    MapImageModule,
+    MapControlsModule,
     MapCenterInfoComponent,
     ModalSearchPageModule,
     ModalMapImagePageModule,
