+++ src/app/modules/shared/shared.module.ts
@@ -23,9 +23,9 @@ import { RefreshWithCancelComponent } from './components/refresh-with-cancel/ref
 import { HeaderComponent } from './components/header/header.component';
 import { OfflineImageComponent } from './components/offline-image/offline-image.component';
 import { CompetenceComponent } from '../../components/competence/competence.component';
-import { MapImageModule } from '../map-image/map-image.module';
 import { AuthModule } from './../auth/auth.module';
 import { VirtualScrollerModule } from 'ngx-virtual-scroller';
+
 @NgModule({
   imports: [
     CommonModule,
@@ -35,7 +35,6 @@ import { VirtualScrollerModule } from 'ngx-virtual-scroller';
     AngularSvgIconModule,
     TranslateModule,
     RouterModule,
-    MapImageModule,
     AuthModule,
     VirtualScrollerModule
   ],
@@ -68,7 +67,6 @@ import { VirtualScrollerModule } from 'ngx-virtual-scroller';
     RouterModule,
     ShadowCssDirective,
     AngularSvgIconModule,
-    MapImageModule,
     AddMenuComponent,
     OfflineImageComponent,
     ExternalLinkComponent,
